fixed value error due to mdmc_reduce
`torchmetrics.functional.f1` calls `torchmetrics.functional.fbeta` which in-turn calls `torchmetrics.functional.stat_score` and it raises value error if inputs are multi-dimensional multi-class and `mdmc_reduce` is not provided. Added `mdmc_average` parameter while defining an instance of the class CustomMetric and it's value could be either `samplewise` or `global`.